Move filter registration from `MainConfig` to `WebConfig`
